requirements.txt
# Django News Portal Requirements

# Core Django
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0

# Database
mysqlclient>=2.1.1  # For MariaDB/MySQL support

# Configuration
python-decouple>=3.6

# Image processing
Pillow>=9.0.0

# Development and testing
pytest-django>=4.5.0
coverage>=6.0.0

# Production server (optional)
gunicorn>=20.1.0

# Static files handling
whitenoise>=6.0.0